pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "datamol"
description = "A python library to work with molecules. Built on top of RDKit."
authors = [{ name = "Hadrien Mary", email = "hadrien@valencediscovery.com" }]
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.8"
license = { text = "Apache" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "tqdm",
    "loguru",
    "joblib",
    "fsspec>=2021.9",
    "pandas",
    "numpy",
    "scipy",
    "matplotlib",
    "pillow",
    "selfies",
    "platformdirs",
    "scikit-learn",
    "packaging",
    "typing-extensions",
    "importlib-resources",
    # NOTE(hadim): can't add rdkit because of `pip` will always override
    # the conda package at the moment.
    # See:
    # - https://github.com/rdkit/rdkit/issues/5378
    # - https://github.com/conda-forge/rdkit-feedstock/issues/104
    # "rdkit",
]

[project.urls]
Website = "https://datamol.io"
"Source Code" = "https://github.com/datamol-io/datamol"
"Bug Tracker" = "https://github.com/datamol-io/datamol/issues"
Documentation = "https://docs.datamol.io"

[tool.setuptools]
include-package-data = true

[tool.setuptools_scm]
fallback_version = "dev"

[tool.setuptools.packages.find]
where = ["."]
include = ["datamol", "datamol.*"]
exclude = []
namespaces = true

[tool.setuptools.package-data]
"datamol.data" = ["*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310']
include = '\.pyi?$'

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--verbose --cov=datamol --cov-fail-under=85 --cov-report xml --cov-report term --durations=10 -n auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:rdkit.*:",
    "ignore::DeprecationWarning:jupyter_client.*:",
    "ignore::DeprecationWarning:pkg_resources.*:",
]

[tool.coverage.run]
source = ["datamol/"]
disable_warnings = ["no-data-collected"]
data_file = ".coverage/coverage"

[tool.coverage.report]
omit = ["datamol/__init__.py", "datamol/_version.py"]

[tool.coverage.xml]
output = "coverage.xml"

[tool.mypy]
exclude = []
ignore_missing_imports = true

[tool.pyright]
reportShadowedImports = false

[tool.ruff]
ignore = [
    "E501", # Never enforce `E501` (line length violations).
    "E731", # Do not assign a lambda expression, use a def
]
line-length = 110
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = [
    "F401", # imported but unused
    "E402", # Module level import not at top of file
]

[tool.ruff.pycodestyle]
max-doc-length = 150
